Migrate MasterServer to WebApplication.CreateBuilder
Drop the legacy WebHost.CreateDefaultBuilder + UseStartup<Startup>
two-file pattern. MasterServerBootstrap now owns the entire bootstrap:
configuration, services, Autofac container, middleware, and endpoint
mapping. Startup.cs deleted.

Behavior preserved: same Serilog/Autofac/Kestrel wiring, same JWT auth,
SignalR hubs, health checks, hosted service, and Mapster setup.
ConfigureServices/InitializeContainer/ConfigurePipeline split kept for
readability.
